Add attribute visitor to LSTMCell op
