+++ src/ngraph/op/fused/lstm_sequence.hpp
@@ -187,22 +187,6 @@ namespace ngraph
         using v0::LSTMSequence;
     } // namespace op
 
-    std::ostream& operator<<(std::ostream& s, const op::LSTMWeightsFormat& type);
-
-    template <>
-    class NGRAPH_API AttributeAdapter<op::LSTMWeightsFormat>
-        : public EnumAttributeAdapterBase<op::LSTMWeightsFormat>
-    {
-    public:
-        AttributeAdapter(op::LSTMWeightsFormat& value)
-            : EnumAttributeAdapterBase<op::LSTMWeightsFormat>(value)
-        {
-        }
-
-        static constexpr DiscreteTypeInfo type_info{"AttributeAdapter<op::LSTMWeightsFormat>", 1};
-        const DiscreteTypeInfo& get_type_info() const override { return type_info; }
-    };
-
     std::ostream& operator<<(std::ostream& s, const op::v0::LSTMSequence::direction& type);
 
     template <>
